Move AppCompatTransferActivity.kt into core/resinjection/transferactivity
